Add a case management map with filters
