+++ corehq/apps/reports/templates/reports/base_template.html
@@ -18,5 +18,6 @@
   {% initial_page_data 'default_config' default_config %}
   {% initial_page_data 'has_datespan' report.has_datespan %}
   {% initial_page_data 'report_configs' report_configs %}
+  {% initial_page_data "mapbox_access_token" mapbox_access_token %}
   {% registerurl 'add_report_config' domain %}
 {% endblock %}
