+++ corehq/apps/reports/templates/reports/standard/base_template.html
@@ -22,6 +22,10 @@
   {% endcompress %}
   {% include 'reports/partials/filters_js.html' %}
   {% include 'reports/partials/graphs/charts_js.html' %}
+    <!-- Scripts for mapbox -->
+  <script src='https://api.mapbox.com/mapbox-gl-js/v2.14.1/mapbox-gl.js'></script>
+  <script src='https://api.mapbox.com/mapbox-gl-js/plugins/mapbox-gl-draw/v1.4.0/mapbox-gl-draw.js'></script>
+  <script src='https://unpkg.com/@turf/turf@6/turf.min.js'></script>
   {% compress js %}
     <script src="{% static 'case/js/cheapxml.js' %}"></script>
     <script src="{% static 'case/js/casexml.js' %}"></script>
@@ -34,12 +38,14 @@
     <script src="{% static 'reports/js/project_health_dashboard.js' %}"></script>
     <script src="{% static 'reports/js/aggregate_user_status.js' %}"></script>
     <script src="{% static 'reports/js/user_history.js' %}"></script>
+    <script src="{% static 'geospatial/js/geospatial_map.js' %}"></script>
   {% endcompress %}
 {% endblock %}
 
 {% block stylesheets %}
   {{ block.super }}
-
+  <link rel="stylesheet" href="https://api.mapbox.com/mapbox-gl-js/v2.14.1/mapbox-gl.css">
+  <link rel="stylesheet" href="https://api.mapbox.com/mapbox-gl-js/plugins/mapbox-gl-draw/v1.4.0/mapbox-gl-draw.css" type="text/css">
   {% compress css %}
     <link type="text/less"
           rel="stylesheet"
